DiffResources: Move versionedResources
